Move config inside internal
